requirements.txt
Babel>=0.9.6
eventlet>=0.12.0
iso8601
oslo.config>=1.2.0
oslo.messaging>=1.3.0
pbr>=0.5.21,<1.0
requests
simplejson
SQLAlchemy>=0.7.8,<0.7.99
sqlalchemy-migrate>=0.8.2
stevedore>=0.14
